% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectToOmlTuneServer.R
\name{connectToOmlTuneServer}
\alias{connectToOmlTuneServer}
\title{Connect to a running OmlTuneServer.}
\usage{
connectToOmlTuneServer(adress = NULL, timeout = 10L)
}
\arguments{
\item{adress}{(\code{character(1)})
The URL where we can reach the API (e.g. \code{http://localhost:8746})}

\item{timeout}{(\code{integer(1)})
How many seconds should we wait until we await a response.
After stating the server it can take a while until it's reachable.}
}
\value{
\code{logical(1)}
}
\description{
Establishes a connection to a runnign OmlTuneServer.
}
