% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnectFromOmlTuneServer.R
\name{disconnectFromOmlTuneServer}
\alias{disconnectFromOmlTuneServer}
\title{Disconnects from a OmlTuneServer.}
\usage{
disconnectFromOmlTuneServer()
}
\value{
\code{logical(1)}
}
\description{
Resets the connection to a OmlTuneServer.
}
