% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllRuns.R
\name{getAllRuns}
\alias{getAllRuns}
\title{Download all runs}
\usage{
getAllRuns(learner.name, task.id)
}
\arguments{
\item{learner.name}{(\code{character(1)})
The learner (e.g. \code{"classif.rpart"})}

\item{task.id}{(\code{integer(1)})
The task id (e.g. \code{3})}
}
\value{
\code{data.frame}
}
\description{
Download all runs
}
