% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAvailableAlgos.R
\name{getAvailableAlgos}
\alias{getAvailableAlgos}
\title{Returns the algo ids for which there are entries in the database.}
\usage{
getAvailableAlgos()
}
\value{
A vector of algo ids.
}
\description{
Returns the algo ids for which there are entries in the database.
}
