% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAvailableTasks.R
\name{getAvailableTasks}
\alias{getAvailableTasks}
\title{Returns the task ids, which are saved in the database and can be queried.}
\usage{
getAvailableTasks()
}
\value{
A vector of task ids.
}
\description{
Returns the task ids, which are saved in the database and can be queried.
}
