% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParamSetForOmlLearner.R
\name{getParamSetForOmlLearner}
\alias{getParamSetForOmlLearner}
\title{Get ParamSet for Learner Name}
\usage{
getParamSetForOmlLearner(learner.name)
}
\arguments{
\item{learner.name}{(\code{character(1)})
The learner class (e.g. "classif.glmnet")}
}
\value{
\code{\link{ParamSet}}
}
\description{
Returns a ParamSet Object for a given Learner
}
