% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOmlBenchFunction.R
\name{makeOmlBenchFunction}
\alias{makeOmlBenchFunction}
\title{Creates a function that will wrap the API}
\usage{
makeOmlBenchFunction(learner.name, task.id, api.chunksize = 20,
  include.extras = FALSE, objective = "accuracy")
}
\arguments{
\item{learner.name}{(\code{character(1)})
Defines the learner. (e.g. \code{"classif.rpart"})}

\item{task.id}{(\code{integer(1)})
Defines the task id the learner is optimized on.}

\item{api.chunksize}{(\code{numeric(1)})
If too many points are requested with one function call this can help to split the request into smaller chunks.}

\item{include.extras}{(\code{logical(1)})
Should the extras be attached as an attribute?}

\item{objective}{(\code{character(1)})
Which value should be optimized? Possible choices are \dQuote{auc}, \dQuote{accuracy} (default) and \dQuote{rmse}.}
}
\value{
\code{function}
}
\description{
Wraps the API Call in a function that returns the performance of a defined learner and \code{task.id} for a given \code{x} using the OmlBench API.
}
