% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startOmlTuneServer.R
\name{startOmlTuneServer}
\alias{startOmlTuneServer}
\title{Start the OmlTuneServer in a docker locally}
\usage{
startOmlTuneServer()
}
\value{
\code{logical(1)}
}
\description{
Start the OmlTuneServer in a docker locally and directly connects to it (assumes that docker is correctly installed).
}
