% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopOmlTuneServer.R
\name{stopOmlTuneServer}
\alias{stopOmlTuneServer}
\title{Stops a locally running OmlTuneServer.}
\usage{
stopOmlTuneServer()
}
\value{
\code{logical(1)}
}
\description{
If you started your OmlTuneServer locally you can stop and disconnect from it with this function.
}
