% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOmlBenchFunction.R
\name{makeOmlBenchFunction}
\alias{makeOmlBenchFunction}
\title{Creates a function that will wrap the API}
\usage{
makeOmlBenchFunction(learner.name, task.id)
}
\arguments{
\item{learner.name}{\code{\link{character(1)}}
Defines the learner.}

\item{task.id}{\code{\link{integer(1)}}
Defines the task id the learner is optimized on.}
}
\value{
\code{function}
}
\description{
Wraps the API Call in a function that returns the performance of a defined learner and \code{task.id} for a given \code{x} using the OmlBench API.
}
